fix: rename table components
