+++ src/components/table/TableHeader.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import PropTypes from '@dhis2/prop-types'
 import { Divider } from '@dhis2/ui'
-import TableRow from '../settings-table/table-row'
+import TableRow from '../settingsTable/TableRow'
 import styles from '../../styles/TableSettings.module.css'
 
 export const TableHeader = ({ title }) => (
