+++ src/pages/Appearance/Datasets/SpecificTableAction.js
@@ -2,7 +2,7 @@ import React, { useState } from 'react'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
 import { datasetHasCategoryCombo } from './helper'
-import TableActions from '../../../components/table-actions'
+import TableActions from '../../../components/TableActions'
 import DialogDelete from '../../../components/dialog/DialogDelete'
 import DialogDatasetSpecificSetting from './DialogDatasetSpecificSetting'
 import {
