+++ src/pages/Synchronization/Datasets/DatasetGlobalSettings.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PropTypes from '@dhis2/prop-types'
-import SettingsTable from '../../../components/settings-table/settings-table'
+import SettingsTable from '../../../components/settingsTable/SettingsTable'
 import { DataSetting } from '../../../constants/data-set-settings'
 import { parseValueBySettingType } from './parseValueBySettingType'
