+++ src/pages/Synchronization/Datasets/SpecificSettings.js
@@ -2,7 +2,7 @@ import React, { useEffect, useState } from 'react'
 import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
 import { Divider } from '@dhis2/ui'
-import TableRow from '../../../components/settings-table/table-row'
+import TableRow from '../../../components/settingsTable/TableRow'
 import Wrapper from '../../../components/wrapper'
 import { InputNumber } from '../../../components/inputs'
 import { DataSpecificSetting } from '../../../constants/data-set-settings'
