+++ src/pages/Synchronization/Datasets/SpecificTableAction.js
@@ -3,7 +3,7 @@ import i18n from '@dhis2/d2-i18n'
 import PropTypes from '@dhis2/prop-types'
 import { parseValueBySettingType } from './parseValueBySettingType'
 import { getPeriodType } from './helper'
-import TableActions from '../../../components/table-actions'
+import TableActions from '../../../components/TableActions'
 import DialogDelete from '../../../components/dialog/DialogDelete'
 import DialogSpecificSettings from './DialogSpecificSettings'
 import {
