+++ src/pages/Synchronization/Programs/ProgramGlobalSettings.js
@@ -9,7 +9,7 @@ import {
     PER_ORG_UNIT,
 } from '../../../constants/program-settings'
 import { SelectSettings } from '../../../components/inputs'
-import SettingsTable from '../../../components/settings-table/settings-table'
+import SettingsTable from '../../../components/settingsTable/SettingsTable'
 import { parseValueByType } from './parseValueBySettingType'
 import { populateProgramObject } from './populateProgramObject'
 import inputStyles from '../../../styles/Input.module.css'
